Move SerializeOptions
